Move eslint config to another directory
