+++ config/eslint/general-rules.js
