+++ eslint.config.js
@@ -2,10 +2,10 @@ const eslintJs = require('@eslint/js');
 const prettier = require('eslint-plugin-prettier');
 const tsEslint = require('typescript-eslint');
 // const eslintJest = require('eslint-plugin-jest');
-const eslintTsRules = require('./conf/eslint-ts-rules');
-const eslintJsRules = require('./conf/eslint-js-rules');
-const eslintReactRules = require('./conf/eslint-react-rules');
-const eslintGeneralRules = require('./conf/eslint-general-rules');
+const eslintTsRules = require('./config/eslint/ts-rules');
+const eslintJsRules = require('./config/eslint/js-rules');
+const eslintReactRules = require('./config/eslint/react-rules');
+const eslintGeneralRules = require('./config/eslint/general-rules');
 const globals = require('globals');
 
 module.exports = tsEslint.config(
